+++ CHANGELOG.md
@@ -8,6 +8,9 @@ CHANGELOG
 
 - Add support for custom actions on login nodes.
 
+**BUG FIXES**
+- Fix validator `EfaPlacementGroupValidator` so that it does not suggest to configure a Placement Group when Capacity Blocks are used.
+
 3.10.1
 ------